# Speed up embedding tests