--- a/tests/models/llm_embed/test_llm_embedding.py
+++ b/tests/models/llm_embed/test_llm_embedding.py
@@ -135,36 +135,36 @@ def test_mpt_embedding_lm(
     lm_config = build_lm_config(is_hf, maybe_attn_impl)
 
     model = ContrastiveModel(**lm_config, tokenizer=mock_tokenizer).to('cuda')
-    # msl = 32
-    # model_inputs_batch = mock_tokenizer([['pair 1 a', 'pair 1 b'],
-    #                                      ['pair 2 a', 'pair 2 b']],
-    #                                     padding='max_length',
-    #                                     truncation=True,
-    #                                     max_length=msl,
-    #                                     return_tensors='pt')
-    # if isinstance(model_inputs_batch, dict):
-    #     model_inputs_batch = {
-    #         k: v.to('cuda') for k, v in model_inputs_batch.items()
-    #     }
+    msl = 32
+    model_inputs_batch = mock_tokenizer([['pair 1 a', 'pair 1 b'],
+                                         ['pair 2 a', 'pair 2 b']],
+                                        padding='max_length',
+                                        truncation=True,
+                                        max_length=msl,
+                                        return_tensors='pt')
+    if isinstance(model_inputs_batch, dict):
+        model_inputs_batch = {
+            k: v.to('cuda') for k, v in model_inputs_batch.items()
+        }
 
-    # with get_precision_context('amp_bf16'):
-    #     outputs = model(model_inputs_batch)
+    with get_precision_context('amp_bf16'):
+        outputs = model(model_inputs_batch)
 
-    #     assert isinstance(outputs, dict)
-    #     assert 'hidden_states' in outputs
+        assert isinstance(outputs, dict)
+        assert 'hidden_states' in outputs
 
-    #     hidden_states = outputs['hidden_states']
-    #     assert isinstance(hidden_states, tuple)
+        hidden_states = outputs['hidden_states']
+        assert isinstance(hidden_states, tuple)
 
-    #     last_hidden_state = hidden_states[-1]
-    #     proj_dim = model.model.config.word_embed_proj_dim
-    #     assert last_hidden_state.shape == (
-    #         4,
-    #         msl,
-    #         proj_dim,
-    #     )  # 2 pairs * 2 texts per pair, 128 sequence length, word_embed_proj_dim dim
-    #     assert last_hidden_state.dtype == torch.bfloat16
-    #     assert last_hidden_state.device.type == 'cuda'
+        last_hidden_state = hidden_states[-1]
+        proj_dim = model.model.config.word_embed_proj_dim
+        assert last_hidden_state.shape == (
+            4,
+            msl,
+            proj_dim,
+        )  # 2 pairs * 2 texts per pair, 128 sequence length, word_embed_proj_dim dim
+        assert last_hidden_state.dtype == torch.bfloat16
+        assert last_hidden_state.device.type == 'cuda'
 
 
 dataloader_config = lambda remote, local_ext: {