--- a/tests/test_utils.py
+++ b/tests/test_utils.py
@@ -142,9 +142,8 @@ class MockTokenizer(PreTrainedTokenizerBase):
         else:
             texts = batch_texts
 
-        max_int = self._vocab_size - max_length
         token_ids = torch.tensor([
-            [hash(text) % max_int + j for j in range(max_length)] for text in texts
+            [hash(text) % 1000 + j for j in range(max_length)] for text in texts
         ])
 
         return {